pyproject.toml
[tool.poetry]
name = "finchat"
version = "0.1.0"
description = ""
authors = ["pkublas <pkublas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.3"
black = "^24.2.0"
boto3 = "^1.34.59"
fastapi = "^0.110.0"
gpt4all = "^2.2.1.post1"
langchain = "^0.1.11"
langchain-openai = "^0.0.8"
openai = "^1.13.3"
numexpr = "^2.9.0"
requests = "^2.31.0"
langchain-community = "^0.0.27"
langchain-core = "^0.1.30"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
